requirements.txt
argon2-cffi==23.1.0
click==8.1.7
Flask==2.3.2
Flask_Login==0.6.3
flask_sqlalchemy==3.1.1
Flask_WTF==1.1.1
Jinja2==3.1.3
markdown2==2.4.13
PyYAML==6.0.2
SQLAlchemy==2.0.30
Werkzeug==2.3.3
WTForms==3.0.1
